backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# File upload
python-multipart==0.0.9

# File storage (AWS S3 for production)
boto3==1.34.69

# Environment
python-dotenv==1.0.1

# Utilities
uuid==1.30